+++ engine/src/main/java/nl/arkenbout/geoffrey/angel/engine/core/graphics/lighting/DirectionalLight.java
@@ -1,41 +1,41 @@
 package nl.arkenbout.geoffrey.angel.engine.core.graphics.lighting;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 import org.lwjgl.util.ReadableColor;
 
 public class DirectionalLight {
 
-    private Vector3f color;
-    private Vector3f direction;
-    private float intensity;
+    private Vector3d color;
+    private Vector3d direction;
+    private double intensity;
 
-    public DirectionalLight(ReadableColor color, Vector3f direction, float intensity) {
-        this(new Vector3f(color.getRed() / 255f, color.getGreen() / 255f, color.getBlue() / 255f), direction, intensity);
+    public DirectionalLight(ReadableColor color, Vector3d direction, double intensity) {
+        this(new Vector3d(color.getRed() / 255f, color.getGreen() / 255f, color.getBlue() / 255f), direction, intensity);
     }
 
-    public DirectionalLight(Vector3f color, Vector3f direction, float intensity) {
+    public DirectionalLight(Vector3d color, Vector3d direction, double intensity) {
         this.color = color;
         this.direction = direction;
         this.intensity = intensity;
     }
 
     public DirectionalLight(DirectionalLight light) {
-        this(new Vector3f(light.getColor()), new Vector3f(light.getDirection()), light.getIntensity());
+        this(new Vector3d(light.getColor()), new Vector3d(light.getDirection()), light.getIntensity());
     }
 
-    public Vector3f getColor() {
+    public Vector3d getColor() {
         return color;
     }
 
-    public Vector3f getDirection() {
+    public Vector3d getDirection() {
         return direction;
     }
 
-    public void setDirection(Vector3f direction) {
+    public void setDirection(Vector3d direction) {
         this.direction = direction;
     }
 
-    public float getIntensity() {
+    public double getIntensity() {
         return intensity;
     }
 }
