+++ engine/src/main/java/nl/arkenbout/geoffrey/angel/engine/core/input/mouse/MouseListener.java
@@ -2,18 +2,17 @@ package nl.arkenbout.geoffrey.angel.engine.core.input.mouse;
 
 import nl.arkenbout.geoffrey.angel.engine.util.Cleanup;
 import org.joml.Vector2d;
-import org.joml.Vector2f;
 
 public interface MouseListener extends Cleanup {
-    void onLeftButtonDown(Vector2f mouseDelta);
+    void onLeftButtonDown(Vector2d mouseDelta);
 
     void onLeftButtonUp();
 
-    void onRightButtonDown(Vector2f mouseDelta);
+    void onRightButtonDown(Vector2d mouseDelta);
 
     void onRightButtonUp();
 
-    void onScroll(float scrollDelta);
+    void onScroll(double scrollDelta);
 
     void onMouseMove(Vector2d mouseDelta);
